requirements.txt
discord.py>=2.3.0
pyodbc>=5.1.0
python-dotenv>=1.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.86.0
google-cloud-pubsub>=2.18.0
aiofiles>=23.1.0
pyperclip>=1.8.2  # for copy button (optional, but we can implement without)